+++ eq_res_partner/models/__init__.py
@@ -23,3 +23,4 @@
 import eq_res_partner
 import eq_res_partner_mails
 import eq_res_company
+import eq_res_users