+++ eq_account/RELEASE_NOTES.md
@@ -1,5 +1,15 @@
 ## Modul eq_account
 
+#### 05.12.2018
+#### Version 1.0.70
+##### IMP
+- Invoice Report: Parent Company name will not be printed anymore when the address is a company as well.
+
+#### 03.12.2018
+#### Version 1.0.69
+##### CHG
+- Account-Report: adding class for future use at report-title
+
 #### 25.10.2018
 #### Version 1.0.68
 ##### FIX